# api/docs/swagger.json
{
    "schemes": [
        "http"
    ], 
    "swagger": "2.0",
    "info": {
        "description": "The public Nebula Labs API for access to pertinent UT Dallas data",
        "title": "nebula-api",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "nebula-api-2lntm5dxoflqn.apigateway.nebula-api-368223.cloud.goog",
    "paths": {
        "/course": {
            "get": {
                "description": "\"Returns all courses matching the query's string-typed key-value pairs\"",
                "produces": [
                    "application/json"
                ],
                "operationId": "courseSearch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The course's official number",
                        "name": "course_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The course's subject prefix",
                        "name": "subject_prefix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The course's title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The course's description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The course's school",
                        "name": "school",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The number of credit hours awarded by successful completion of the course",
                        "name": "credit_hours",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The level of education that this course course corresponds to",
                        "name": "class_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The type of class this course corresponds to",
                        "name": "activity_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The grading status of this course",
                        "name": "grading",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The internal (university) number used to reference this course",
                        "name": "internal_course_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The weekly contact hours in lecture for a course",
                        "name": "lecture_contact_hours",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The frequency of offering a course",
                        "name": "offering_frequency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of courses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Course"
                            }
                        }
                    }
                }
            }
        },
        "/course/{id}": {
            "get": {
                "description": "\"Returns the course with given ID\"",
                "produces": [
                    "application/json"
                ],
                "operationId": "courseById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the course to get",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A course",
                        "schema": {
                            "$ref": "#/definitions/schema.Course"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.CollectionRequirement": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {}
                },
                "required": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "schema.Course": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "activity_type": {
                    "type": "string"
                },
                "attributes": {},
                "catalog_year": {
                    "type": "string"
                },
                "class_level": {
                    "type": "string"
                },
                "co_or_pre_requisites": {
                    "$ref": "#/definitions/schema.CollectionRequirement"
                },
                "corequisites": {
                    "$ref": "#/definitions/schema.CollectionRequirement"
                },
                "course_number": {
                    "type": "string"
                },
                "credit_hours": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enrollment_reqs": {
                    "type": "string"
                },
                "grading": {
                    "type": "string"
                },
                "internal_course_number": {
                    "type": "string"
                },
                "laboratory_contact_hours": {
                    "type": "string"
                },
                "lecture_contact_hours": {
                    "type": "string"
                },
                "offering_frequency": {
                    "type": "string"
                },
                "prerequisites": {
                    "$ref": "#/definitions/schema.CollectionRequirement"
                },
                "school": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subject_prefix": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    },
...
}